docker/base/ccapp/requirements.txt
Django==5.1
gunicorn==20.1
psycopg2==2.9
djangorestframework==3.15
djangorestframework-simplejwt==5.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
attrs
django-cors-headers==4.6
bcrypt==4.2
channels==4.2.2
daphne==4.1.2
channels_redis==4.2.1
twisted[http2,tls]==24.11.0
uvicorn[standard]==0.34.0
